fix: move homepage state context related files closer
