+++ packages/apps/human-app/frontend/src/modules/homepage/components/choose-sign-up-account-type.tsx
@@ -8,7 +8,7 @@ import { PageCard } from '@/shared/components/ui/page-card';
 import type { HomePageStageType } from '@/modules/homepage/views/home.page';
 import { useColorMode } from '@/shared/hooks/use-color-mode';
 import { onlyDarkModeColor } from '@/shared/styles/dark-color-palette';
-import { useHomePageState } from '@/shared/hooks/use-homepage-state';
+import { useHomePageState } from '@/shared/contexts/homepage-state';
 
 interface ChooseSignUpAccountType {
   setStage: (step: HomePageStageType) => void;
