+++ packages/apps/human-app/frontend/src/shared/components/layout/unprotected/navbar.tsx
@@ -13,7 +13,7 @@ import { breakpoints } from '@/shared/styles/breakpoints';
 import { env } from '@/shared/env';
 import { DarkModeSwitch } from '@/shared/components/ui/dark-mode-switch';
 import { useHandleMainNavIconClick } from '@/shared/hooks/use-handle-main-nav-icon-click';
-import { useHomePageState } from '@/shared/hooks/use-homepage-state';
+import { useHomePageState } from '@/shared/contexts/homepage-state';
 
 interface NavbarProps {
   withNavigation: boolean;
