--- packages/apps/human-app/frontend/src/shared/types/homepage-state.ts
@@ -1,1 +1,0 @@
-export type HomePageStateType = 'welcome' | 'chooseSignUpAccountType';
